Validation

+++ src/main/java/com/aaludra/spring/jpa/h2/repository/EmployeeRepository.java
@@ -1,14 +1,14 @@
 package com.aaludra.spring.jpa.h2.repository;
 
-import java.sql.Date;
+
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aaludra.spring.jpa.h2.model.Employee;
-import com.aaludra.spring.jpa.h2.model.Tutorial;
 
-public interface EmployeeRepository extends JpaRepository <Employee, Integer>{
+
+public interface EmployeeRepository extends JpaRepository <Employee, Long>{
 
 	 List<Employee> findByempName(String empName);
